+++ be/src/main/java/com/codesquad/issuetracker/api/member/Service/MemberService.java
@@ -1,15 +1,15 @@
 package com.codesquad.issuetracker.api.member.Service;
 
 import com.codesquad.issuetracker.api.member.domain.Member;
-import com.codesquad.issuetracker.api.member.dto.SignInRequest;
-import com.codesquad.issuetracker.api.member.dto.SignUpRequest;
+import com.codesquad.issuetracker.api.member.dto.request.SignInRequest;
+import com.codesquad.issuetracker.api.member.dto.request.SignUpRequest;
+import com.codesquad.issuetracker.api.member.dto.response.SignInResponse;
 import com.codesquad.issuetracker.api.member.repository.MemberRepository;
 import com.codesquad.issuetracker.api.member.repository.TokenRepository;
-import com.codesquad.issuetracker.api.oauth.dto.SignInResponse;
-import com.codesquad.issuetracker.api.oauth.dto.UserProfile;
-import com.codesquad.issuetracker.api.oauth.jwt.Jwt;
-import com.codesquad.issuetracker.api.oauth.jwt.JwtProvider;
-import com.codesquad.issuetracker.api.oauth.service.OauthService;
+import com.codesquad.issuetracker.jwt.Jwt;
+import com.codesquad.issuetracker.jwt.JwtProvider;
+import com.codesquad.issuetracker.oauth.dto.response.OauthUserProfile;
+import com.codesquad.issuetracker.oauth.service.OauthService;
 import java.util.Map;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
@@ -26,8 +26,8 @@ public class MemberService {
 
     public SignInResponse oAuthSignIn(String providerName, String code) {
         //github에서 사용자 정보 가져오기
-        UserProfile userProfile = oauthService.login(providerName, code);
-        Member member = userProfile.toEntity();
+        OauthUserProfile oauthUserProfile = oauthService.login(providerName, code);
+        Member member = oauthUserProfile.toEntity();
 
         Optional<Long> memberId = memberRepository.findBy(member.getEmail());
         if (!memberId.isPresent()) {
