+++ be/src/main/java/com/codesquad/issuetracker/oauth/config/OauthConfig.java
@@ -1,9 +1,6 @@
-package com.codesquad.issuetracker.api.oauth.config;
+package com.codesquad.issuetracker.oauth.config;
 
-import com.codesquad.issuetracker.api.oauth.InMemoryProviderRepository;
-import com.codesquad.issuetracker.api.oauth.OauthAdapter;
-import com.codesquad.issuetracker.api.oauth.OauthProperties;
-import com.codesquad.issuetracker.api.oauth.OauthProvider;
+import com.codesquad.issuetracker.oauth.domain.OauthProvider;
 import java.util.Map;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
