+++ be/src/main/java/com/codesquad/issuetracker/oauth/dto/response/OauthUserProfile.java
@@ -1,17 +1,17 @@
-package com.codesquad.issuetracker.api.oauth.dto;
+package com.codesquad.issuetracker.oauth.dto.response;
 
 import com.codesquad.issuetracker.api.member.domain.Member;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
-public class UserProfile {
+public class OauthUserProfile {
     private final String email;
     private final String name;
     private final String imageUrl;
 
     @Builder
-    public UserProfile(String email, String name, String imageUrl) {
+    public OauthUserProfile(String email, String name, String imageUrl) {
         this.email = email;
         this.name = name;
         this.imageUrl = imageUrl;
