+++ be/src/main/java/com/codesquad/issuetracker/oauth/service/OauthService.java
@@ -1,10 +1,10 @@
-package com.codesquad.issuetracker.api.oauth.service;
+package com.codesquad.issuetracker.oauth.service;
 
-import com.codesquad.issuetracker.api.oauth.InMemoryProviderRepository;
-import com.codesquad.issuetracker.api.oauth.OauthAttributes;
-import com.codesquad.issuetracker.api.oauth.OauthProvider;
-import com.codesquad.issuetracker.api.oauth.dto.OauthTokenResponse;
-import com.codesquad.issuetracker.api.oauth.dto.UserProfile;
+import com.codesquad.issuetracker.oauth.OauthAttributes;
+import com.codesquad.issuetracker.oauth.config.InMemoryProviderRepository;
+import com.codesquad.issuetracker.oauth.domain.OauthProvider;
+import com.codesquad.issuetracker.oauth.dto.response.OauthTokenResponse;
+import com.codesquad.issuetracker.oauth.dto.response.OauthUserProfile;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.Map;
@@ -22,7 +22,7 @@ public class OauthService {
 
     private final InMemoryProviderRepository inMemoryProviderRepository;
 
-    public UserProfile login(String providerName, String code) {
+    public OauthUserProfile login(String providerName, String code) {
         // 프론트에서 넘어온 provider 이름을 통해 InMemoryProviderRepository에서 OauthProvider 가져오기
         OauthProvider provider = inMemoryProviderRepository.findByProviderName(providerName);
 
@@ -30,9 +30,9 @@ public class OauthService {
         OauthTokenResponse tokenResponse = getToken(code, provider);
 
         // 유저 정보 가져오기
-        UserProfile userProfile = getUserProfile(providerName, tokenResponse, provider);
+        OauthUserProfile oauthUserProfile = getUserProfile(providerName, tokenResponse, provider);
 
-        return userProfile;
+        return oauthUserProfile;
     }
 
     private OauthTokenResponse getToken(String code, OauthProvider provider) {
@@ -59,7 +59,8 @@ public class OauthService {
         return formData;
     }
 
-    private UserProfile getUserProfile(String providerName, OauthTokenResponse tokenResponse, OauthProvider provider) {
+    private OauthUserProfile getUserProfile(String providerName, OauthTokenResponse tokenResponse,
+                                            OauthProvider provider) {
         Map<String, Object> userAttributes = getUserAttributes(provider, tokenResponse);
         return OauthAttributes.extract(providerName, userAttributes);
     }
